New bindings

+++ src/constants/RTE_MBUF.rs
@@ -2,7 +2,7 @@
 // Copyright © 2016 The developers of dpdk-sys. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/dpdk-sys/master/COPYRIGHT.
 
 
-pub const RTE_MBUF_DEFAULT_BUF_SIZE: c_int = 2176;
+pub const RTE_MBUF_DEFAULT_BUF_SIZE: uint16_t = 2176;
 pub const RTE_MBUF_DEFAULT_DATAROOM: c_int = 2048;
 pub const RTE_MBUF_PRIV_ALIGN: c_int = 8;
 pub const RTE_MBUF_REFCNT_ATOMIC: c_int = 1;
